activemq-private-lambda-python-sam/activemq_message_sender_python/requirements.txt
stomp.py>=8.1.0
boto3>=1.26.0
